proje bitiş

+++ BLL/ConcreteService/GuestService.cs
@@ -15,20 +15,35 @@ namespace BLL.ConcreteService
     {
         private readonly IGenericRepository<Guest> _genericRepository;
         private readonly IMapper _mapper;
+        private readonly IPostService _postService;
 
-        public GuestService(IGenericRepository<Guest> genericRepository,IMapper mapper)
+        public GuestService(IGenericRepository<Guest> genericRepository,IMapper mapper, IPostService postService)
         {
             _genericRepository = genericRepository;
             _mapper = mapper;
+            _postService = postService;
         }
         public async Task AddGuestAsync(GuestDto guestDto)
         {
             await _genericRepository.AddAsync(_mapper.Map<Guest>(guestDto));
         }
 
-        public async Task DeleteGuestAsync(int Id)
+        public async Task<bool> DeleteGuestAsync(int Id)
         {
+            var guest = await _genericRepository.GetByIdAsync(Id);
+            if (guest == null)
+            {
+                return false; // Kullanıcı bulunamadı
+            }
+
+            var posts = await _postService.GetPostsByGuestIdAsync(Id);
+            foreach (var post in posts)
+            {
+                await _postService.DeletePostAsync(post.Id);
+            }
+
             await _genericRepository.DeleteAsync(Id);
+            return true;
         }
 
         public async Task<List<GuestDto>> GetAllGuestAsync()
@@ -44,7 +59,8 @@ namespace BLL.ConcreteService
 
         public async Task UpdateGuestAsync(GuestDto guestDto, int Id)
         {
-            await _genericRepository.UpdateAsync(_mapper.Map<Guest>(guestDto), Id);
+            var guest = _mapper.Map<Guest>(guestDto);
+            await _genericRepository.UpdateAsync(guest, Id);
         }
     }
 }
