+++ BLL/ConcreteService/PostService.cs
@@ -3,6 +3,7 @@ using BLL.AbstractService;
 using BLL.AllDto;
 using DAL.AbstractRepository;
 using DAL.Entities;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -16,7 +17,7 @@ namespace BLL.ConcreteService
         private readonly IGenericRepository<Post> _genericRepository;
         private readonly IMapper _mapper;
 
-        public PostService(IGenericRepository<Post> genericRepository,IMapper mapper)
+        public PostService(IGenericRepository<Post> genericRepository, IMapper mapper)
         {
             _genericRepository = genericRepository;
             _mapper = mapper;
@@ -28,17 +29,42 @@ namespace BLL.ConcreteService
 
         public async Task DeletePostAsync(int Id)
         {
-            await _genericRepository.DeleteAsync(Id);
+            var post = await _genericRepository.GetByIdAsync(Id);
+            if (post != null)
+            {
+                await _genericRepository.DeleteAsync(Id);
+            }
         }
 
         public async Task<List<PostDto>> GetAllPostAsync()
         {
-            return _mapper.Map<List<PostDto>>(await _genericRepository.GetAllAsync());    
+            return _mapper.Map<List<PostDto>>(await _genericRepository.GetAllAsync());
         }
 
         public async Task UpdatepostAsync(PostDto postDto, int Id)
         {
             await _genericRepository.UpdateAsync(_mapper.Map<Post>(postDto), Id);
         }
+        public async Task<List<Post>> GetPostsByGuestIdAsync(int guestId)
+        {
+            var posts = await _genericRepository.GetAllAsync();
+            return posts.Where(p => p.GuestId == guestId).ToList();
+        }
+
+        public async Task<PostDto> GetPostByIdAsync(int Id)
+        {
+            var post = await _genericRepository.GetByIdAsync(Id);
+            if (post == null)
+            {
+                return null; 
+            }
+
+            return _mapper.Map<PostDto>(post);
+
+        }
+
+       
+
     }
 }
+
